changed key method to take 8 bit int

+++ header.h
@@ -6,7 +6,7 @@ void encrypt(char* txtFile);
 void decrypt(char* binFile);
 void promptUser(void);
 int getSeed(char* fileName);
-int getKey(int currentKey);
+int8_t getKey(int8_t currentKey);
 char* readFileName(void);
 
 #endif
